// types.ts
export interface ReportCardType {
  title: string;
  description: string;
  icon: React.ElementType;
  category: 'Financial Statements' | 'Sales & Receivables' | 'Purchases & Payables' | 'Tax';
}